--- src/app/login/page.tsx
@@ -1,5 +1,0 @@
-import LoginPageComponent from '@/components/login-page'
-
-export default async function LoginPage() {
-  return <LoginPageComponent />
-}
