+++ src/app/sign-up/page.tsx
@@ -1,0 +1,5 @@
+import RegisterPageComponent from '@/components/register-page'
+
+export default async function LoginPage() {
+  return <RegisterPageComponent />
+}
